Revert "Add watchdog timer & [Rename package]"
