+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.8)
-project(bno08x_driver)
+project(bno08x_ros2_driver)
 
 # Check for compiler options
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
@@ -78,12 +78,10 @@ if(BUILD_TESTING)
     target_include_directories(test_i2c_mock PUBLIC
       $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
     )
-  endif()
-
-  ament_add_gmock(test_watchdog test/watchdog_test.cpp)
-  if(TARGET test_watchdog)
-    target_include_directories(test_watchdog PUBLIC
-      $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    ament_target_dependencies(test_i2c_mock
+      "rclcpp"
+      "sensor_msgs"
+      "std_msgs"
     )
   endif()
